refactor: add Concurrency and Permissions structures for workflow management
