+++ crates/gh-workflow/src/workflow.rs
@@ -9,9 +9,11 @@ use merge::Merge;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
+use crate::concurrency::Concurrency;
 use crate::error::Result;
 use crate::generate::Generate;
 use crate::job::Job;
+use crate::permissions::Permissions;
 use crate::{Event};
 
 #[derive(Debug, Default, Serialize, Deserialize, Clone)]
@@ -304,85 +306,9 @@ impl Volume {
     }
 }
 
-/// Represents concurrency settings for workflows.
-#[derive(Debug, Setters, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
-#[serde(rename_all = "kebab-case")]
-#[setters(strip_option, into)]
-pub struct Concurrency {
-    /// The group name for concurrency.
-    pub group: String,
 
-    /// Whether to cancel in-progress jobs.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cancel_in_progress: Option<bool>,
 
-    /// The limit on concurrent jobs.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub limit: Option<u32>,
-}
 
-impl Concurrency {
-    pub fn new(group: impl Into<Expression>) -> Self {
-        let expr: Expression = group.into();
-        Self { group: expr.0, ..Default::default() }
-    }
-}
-
-/// Represents permissions for the `GITHUB_TOKEN`.
-#[derive(Debug, Setters, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
-#[serde(rename_all = "kebab-case")]
-#[setters(strip_option, into)]
-pub struct Permissions {
-    /// Permissions for actions.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub actions: Option<Level>,
-
-    /// Permissions for repository contents.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub contents: Option<Level>,
-
-    /// Permissions for issues.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub issues: Option<Level>,
-
-    /// Permissions for pull requests.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pull_requests: Option<Level>,
-
-    /// Permissions for deployments.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub deployments: Option<Level>,
-
-    /// Permissions for checks.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub checks: Option<Level>,
-
-    /// Permissions for statuses.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub statuses: Option<Level>,
-
-    /// Permissions for packages.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub packages: Option<Level>,
-
-    /// Permissions for pages.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pages: Option<Level>,
-
-    /// Permissions for ID tokens.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id_token: Option<Level>,
-}
-
-/// Represents the level of permissions.
-#[derive(Debug, Serialize, Deserialize, Clone, Copy, Default, PartialEq, Eq)]
-#[serde(rename_all = "kebab-case")]
-pub enum Level {
-    Read,
-    Write,
-    #[default]
-    None,
-}
 
 /// Represents the strategy for running jobs.
 #[derive(Debug, Setters, Serialize, Deserialize, Clone, Default, PartialEq, Eq)]
